requirements.txt
scikit-learn==0.24.0
numpy==1.19.2
pandas==1.1.5
transformers==4.1.1
networkx==2.5
scipy==1.5.3
tqdm==4.54.1
python==3.8.5
pytorch==1.7.1